Upgraded Bukkit from 1.12.2 to 1.14.4
The version of bukkit used by rasberry juice is quite old and when I try to start bukkit with it installed I get an error:

    org.bukkit.plugin.InvalidPluginException: java.lang.NoSuchFieldError: GOLD_SWORD

I think that in version 1.13 of bukkit, `GOLD_SWORD` and `WOOD_SWORD` were renamed to `GOLDEN_SWORD` and `WOODEN_SWORD`, and although that doesn't stop rasberry juice from compiling, the compiled code is incompatible with up-to-date bukkit servers.

To fix that I have updated the pom file to compile against bukkit 1.14.4 (the latest on the spigot repository). As you would expect, that produced compile errors at the same place I'm seeing the issue in the bukkit logs:

    [ERROR] RaspberryJuice/src/main/java/net/zhuoweizhang/raspberryjuice/RaspberryJuicePlugin.java:[21,11] error: cannot find symbol
    [ERROR]
    [ERROR] could not parse error message:   symbol:   variable GOLD_SWORD
    [ERROR]   location: class Material
    [ERROR] RaspberryJuice/src/main/java/net/zhuoweizhang/raspberryjuice/RaspberryJuicePlugin.java:24: error: cannot find symbol
    [ERROR] 			Material.WOOD_SWORD);
    [ERROR] 			        ^

Updating those to the current variable names fixes the compile error, and the compiled plugin now works with bukkit.